fix: rename welcome pages and flag to progressive profiling
