+++ .env
@@ -18,11 +18,11 @@ AUTHN_MINIMAL_HEADER=true
 LOGIN_ISSUE_SUPPORT_LINK=''
 USER_SURVEY_COOKIE_NAME=null
 COOKIE_DOMAIN=null
-WELCOME_PAGE_SUPPORT_LINK=null
+AUTHN_PROGRESSIVE_PROFILING_SUPPORT_LINK=null
 INFO_EMAIL=''
 DISABLE_ENTERPRISE_LOGIN=''
 REGISTER_CONVERSION_COOKIE_NAME=null
-ENABLE_PROGRESSIVE_PROFILING=''
+ENABLE_PROGRESSIVE_PROFILING_ON_AUTHN=''
 MARKETING_EMAILS_OPT_IN=''
 ENABLE_COPPA_COMPLIANCE=''
 SHOW_CONFIGURABLE_EDX_FIELDS=''
