+++ .env.test
@@ -19,7 +19,7 @@ SITE_NAME='Your Platform Name Here'
 USER_INFO_COOKIE_NAME='edx-user-info'
 LOGIN_ISSUE_SUPPORT_LINK='https://login-issue-support-url.com'
 USER_SURVEY_COOKIE_NAME='openedx-user-survey-type'
-WELCOME_PAGE_SUPPORT_LINK='http://localhost:1999/welcome'
+AUTHN_PROGRESSIVE_PROFILING_SUPPORT_LINK='http://localhost:1999/welcome'
 DISABLE_ENTERPRISE_LOGIN=''
 REGISTER_CONVERSION_COOKIE_NAME='openedx-user-register-conversion'
 MARKETING_EMAILS_OPT_IN=''
