+++ src/progressive-profiling/data/reducers.js
