+++ src/progressive-profiling/data/selectors.js
